There were 2 productSlice.js, removes bad file. Renames goods folder to products and updats imports
