+++ src/features/products/DetailedProductCard.js
