+++ src/features/products/productSlice.js
@@ -1,4 +1,4 @@
-import { PRODUCTS } from "./PRODUCTS";
+import { PRODUCTS } from '../../shared/PRODUCTS';
 
 //selector for Type -- given type as argument, this function should return the products with that type from the PRODUCTS array
 export const selectProductsByType = (type) => {
@@ -7,9 +7,9 @@ export const selectProductsByType = (type) => {
   );
 };
 
-//selector for Top / Category of Products -- this function returns the categories of products for the Products Page
-export const selectTopProducts = () => {
-  return PRODUCTS.filter((product) => product.type === "top");
+//selector for Category of Products -- this function returns the categories of products for the Products Page
+export const selectCategories = () => {
+  return PRODUCTS.filter((product) => product.type === 'category');
 };
 
 //selector for Seasonal Products -- this function returns Seasonal Products only
